+++ NEWS.rst
@@ -6,6 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
-* Major: Upgrade to Kodo 30
-* Major: Renamed ``decode`` to ``read_payload`` and ``encode`` and ``recode`` to ``write_payload``
+* Major: Upgrade to waf-tools 3
+* Major: Upgrade to kodo 32
+* Major: Renamed ``set_symbol`` to ``set_const_symbol``.
+* Major: Renamed ``set_symbols`` to ``set_const_symbols``.
+* Major: Renamed ``decode`` to ``read_payload`` and ``encode`` and ``recode``
+  to ``write_payload``
 * Major: Renamed ``copy_symbols`` to ``copy_from_symbols``